requirements.txt
docx2txt==0.8
langchain==0.0.294
openai==0.28.0
PyPDF2==3.0.1
python-pptx==0.6.22
streamlit==1.27.0
